evaluation/Model_Centric/AdaptVis/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
clip==0.2.0
comm==0.2.3
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
easydict==1.13
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.10
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.2
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
setuptools==78.1.1
six==1.17.0
stack_data==0.6.3
sympy==1.13.3
tokenizers==0.15.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.39.1
triton==3.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
